back/reaction/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework
djongo
mongoengine==0.8.0
requests==2.25.1